% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_data_gov_url.R
\name{all_data_gov_url}
\alias{all_data_gov_url}
\title{Check if all URLs are from data.gov.il}
\usage{
all_data_gov_url(url)
}
\arguments{
\item{url}{Character vector of URLs to check}
}
\value{
Logical scalar indicating if all URLs are from data.gov.il
}
\description{
Check if all URLs are from data.gov.il
}
\examples{
 @noRd
}
