% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gidi_org.R
\name{as_gidi_org}
\alias{as_gidi_org}
\title{Create a new organization object}
\usage{
as_gidi_org(x)
}
\arguments{
\item{x}{A named list containing organization data}
}
\value{
A gidi_org or gidi_org_full_info object
}
\description{
Create a new organization object
}
\examples{
org_data <- list(
  id = "123",
  eng_name = "Test Org",
  heb_name = "ארגון בדיקה"
)
as_gidi_org(org_data)
}
