% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gidi_resource2.R
\name{as_gidi_resource}
\alias{as_gidi_resource}
\title{Convert a Named List to a gidi_resource Object}
\usage{
as_gidi_resource(x)
}
\arguments{
\item{x}{A named list with fields `resource_id`, `pak_heb_name`,
`pak_eng_name`, `resource_heb_name`, `size`, `last_modified`, and
`created`.}
}
\value{
A [gidi_resource] object.
}
\description{
Convert a Named List to a gidi_resource Object
}
\seealso{
Other gidi objects: 
\code{\link{as.data.frame.gidi}},
\code{\link{as_gidi_org}()},
\code{\link{as_gidi_pak}()},
\code{\link{gidi_obj}()},
\code{\link{gidi_obj_class}},
\code{\link{gidi_objs_list}()},
\code{\link{gidi_org_full_info}()},
\code{\link{gidi_resource}()},
\code{\link{list_gidi_org}()},
\code{\link{list_gidi_pak}()},
\code{\link{list_gidi_resource}()}
}
\concept{gidi objects}
