% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gidi_city_code.R
\name{city_code_req}
\alias{city_code_req}
\title{Create an API request to CBS localities endpoint}
\usage{
city_code_req(url_query)
}
\arguments{
\item{url_query}{List of query parameters}
}
\value{
An httr2 request object
}
\description{
Create an API request to CBS localities endpoint
}
\keyword{internal}
