% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_utils.R
\name{class_checker}
\alias{class_checker}
\title{Generic: build_class_check}
\usage{
class_checker(val, class, ...)
}
\description{
Returns a function that checks if a value inherits
from the specified class.
}
