% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gidi_city_code.R
\name{extract_df_city_code}
\alias{extract_df_city_code}
\title{Extract and format locality data from API response}
\usage{
extract_df_city_code(x)
}
\arguments{
\item{x}{Raw API response data}
}
\value{
A data frame with city codes and names
}
\description{
Extract and format locality data from API response
}
\keyword{internal}
