% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gidigov.R
\name{get_package_info}
\alias{get_package_info}
\title{Get package information}
\usage{
get_package_info(package_id)
}
\arguments{
\item{package_id}{Character string of package ID}
}
\value{
data.table of active resources in the package
}
\description{
Get package information
}
