% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gidigov.R
\name{get_resource_id_by_name}
\alias{get_resource_id_by_name}
\title{Get resource ID by name}
\usage{
get_resource_id_by_name(database_name)
}
\arguments{
\item{database_name}{Character string of database name}
}
\value{
Character string of resource ID
}
\description{
Get resource ID by name
}
