% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gidigov.R
\name{get_resource_info}
\alias{get_resource_info}
\title{Create an offset request}
\usage{
get_resource_info(resource_id)
}
\arguments{
\item{resource_id}{Character vector of resource IDs}

\item{url}{Base URL}

\item{offset}{Offset value for pagination}
}
\value{
An httr2 request object

A data.table containing resource information
}
\description{
Create an offset request
}
