% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gidigov.R
\name{get_resource_name}
\alias{get_resource_name}
\title{Get resource name by ID}
\usage{
get_resource_name(resource_id)
}
\arguments{
\item{resource_id}{Character string of resource ID}
}
\value{
Character string of resource name
}
\description{
Get resource name by ID
}
