% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gidi_obj_class.R
\name{gidi_obj_class}
\alias{gidi_obj_class}
\title{Base Classes for gidi Objects}
\description{
Abstract base classes for the gidigov object hierarchy:
- [gidi_obj]: Abstract base for individual entities (organizations,
  packages, resources).
- [gidi_objs_list]: Abstract base for typed lists of gidi objects.

All gidi objects support [as.data.frame()] and pretty-print methods.
}
\seealso{
Other gidi objects: 
\code{\link{as.data.frame.gidi}},
\code{\link{as_gidi_org}()},
\code{\link{as_gidi_pak}()},
\code{\link{as_gidi_resource}()},
\code{\link{gidi_obj}()},
\code{\link{gidi_objs_list}()},
\code{\link{gidi_org_full_info}()},
\code{\link{gidi_resource}()},
\code{\link{list_gidi_org}()},
\code{\link{list_gidi_pak}()},
\code{\link{list_gidi_resource}()}
}
\concept{gidi objects}
