% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gidi_resource2.R
\name{gidi_resource}
\alias{gidi_resource}
\title{Base class for resource representation}
\usage{
gidi_resource(
  resource_id = character(0),
  pak_heb_name = character(0),
  pak_eng_name = character(0),
  resource_heb_name = character(0),
  size = integer(0),
  last_modified = (function (.data = double()) 
 {
     .Date(.data)
 })(),
  created = (function (.data = double()) 
 {
     .Date(.data)
 })()
)
}
\description{
Creates a new resource class for the Israeli government data repository
}
