% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_data_gov_url.R
\name{is_data_gov_url}
\alias{is_data_gov_url}
\title{Check if URL is from data.gov.il}
\usage{
is_data_gov_url(url)
}
\arguments{
\item{url}{Character vector of URLs to check}
}
\value{
Logical vector indicating if each URL starts with "https://data.gov.il/"
}
\description{
Check if URL is from data.gov.il
}
\examples{
is_data_gov_url("https://data.gov.il/dataset/123")
 @noRd
}
