% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gidi_org.R
\name{list_gidi_org}
\alias{list_gidi_org}
\title{List class for holding multiple organizations}
\usage{
list_gidi_org(.data = list())
}
\description{
Creates a list class for managing multiple organization objects
}
