% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gidi_resource2.R
\name{list_gidi_resource}
\alias{list_gidi_resource}
\title{List class for holding multiple resources}
\usage{
list_gidi_resource(.data = list())
}
\description{
Creates a list class for managing multiple resource objects
}
