% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gidi_city_code.R
\name{modify_queries_pages}
\alias{modify_queries_pages}
\title{Generate query parameters for additional pages}
\usage{
modify_queries_pages(resp_info, base_query)
}
\arguments{
\item{resp_info}{Response info containing pagination data}

\item{base_query}{Original query parameters}
}
\value{
List of query parameters for remaining pages
}
\description{
Generate query parameters for additional pages
}
\keyword{internal}
