% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_list_of.R
\name{new_class_list_of}
\alias{new_class_list_of}
\title{Create a new S7 class for lists of a specific class}
\usage{
new_class_list_of(class, allow_null = FALSE, names = c("any", "all", "none"))
}
\arguments{
\item{class}{An S7 class object, a base class, a union of classes,
or a character string giving the name of a class.}

\item{allow_null}{Logical; whether `NULL` is allowed for the list.}

\item{names}{A character string indicating naming requirements for
list elements. One of `"any"`, `"all"`, or `"none"`.}
}
\value{
An S7 class object with built-in validation rules.
}
\description{
Defines an S7 class for a list whose elements inherit
from a given class. Includes validation for `NULL` values
and naming requirements.
}
