% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gidi_obj_class.R
\name{new_gidi_objs_list}
\alias{new_gidi_objs_list}
\title{Create a new gidi objects list class}
\usage{
new_gidi_objs_list(gidi_obj_class)
}
\arguments{
\item{gidi_obj_class}{The gidi object class that elements of the list must inherit from}
}
\value{
A new S7 class for lists of the specified gidi object type
}
\description{
Factory function to create a new class for lists of specific gidi objects.
The new class inherits from gidi_objs_list and includes validation to ensure
all elements are of the specified gidi object class.
}
\examples{
# Create a new class for lists of gidi_pak objects
list_gidi_pak <- new_gidi_objs_list(gidi_pak)
}
