% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gidi_city_code.R
\name{update_page_number}
\alias{update_page_number}
\title{Update page number in query parameters}
\usage{
update_page_number(page_number, base_params)
}
\arguments{
\item{page_number}{Page number to set}

\item{base_params}{Original query parameters}
}
\value{
Updated query parameters list
}
\description{
Update page number in query parameters
}
\keyword{internal}
