% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gidi_city_code.R
\name{update_page_numbers}
\alias{update_page_numbers}
\title{Vectorized version of update_page_number}
\usage{
update_page_numbers(page_number, base_params)
}
\description{
Vectorized version of update_page_number
}
\keyword{internal}
