% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_data_gov_url.R
\name{url_to_id}
\alias{url_to_id}
\title{Convert URL to ID based on type}
\usage{
url_to_id(url, type)
}
\arguments{
\item{url}{Character vector of data.gov.il URLs}

\item{type}{Character scalar of expected URL type
(one of: "resource", "organization", "dataset")}
}
\value{
Character vector of IDs if URLs match type, otherwise original URLs
 @noRd
}
\description{
Convert URL to ID based on type
}
