% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gidi_obj_class.R
\name{as.data.frame.gidi}
\alias{as.data.frame.gidi}
\title{Convert gidi objects to data frames}
\arguments{
\item{x}{A gidi_obj or gidi_objs_list object}

\item{...}{Additional arguments passed to as.data.frame}
}
\value{
A data frame representation of the object
}
\description{
Convert gidi objects to data frames
}
