% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gidi_pak.R
\name{as_gidi_pak}
\alias{as_gidi_pak}
\title{#' Print method for pak list
#' @param x A list_gidi_pak object
#' @param ... Additional arguments passed to print
S7::method(print, list_gidi_pak) <- function(x, ...) {
  print(S7::S7_data(x))
}
Create a new pak object}
\usage{
as_gidi_pak(x)
}
\arguments{
\item{x}{A named list containing pak data}
}
\value{
A gidi_pak object
}
\description{
#' Print method for pak list
#' @param x A list_gidi_pak object
#' @param ... Additional arguments passed to print
S7::method(print, list_gidi_pak) <- function(x, ...) {
  print(S7::S7_data(x))
}
Create a new pak object
}
\examples{
pak_data <- list(
  pak_id = "12345",
  pak_heb_name = "חבילה לדוגמה",
  pak_eng_name = "Example Package",
  org_heb_name = "ארגון לדוגמה",
  org_eng_name = "Example Organization"
)
as_gidi_pak(pak_data)
}
