% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gidi_resource2.R
\name{as_gidi_resource}
\alias{as_gidi_resource}
\title{Create a new resource object}
\usage{
as_gidi_resource(x)
}
\arguments{
\item{x}{A named list containing resource data}
}
\value{
A gidi_resource object
}
\description{
Create a new resource object
}
\examples{
resource_data <- list(
  resource_id = "56789",
  pak_heb_name = "חבילה לדוגמה",
  pak_eng_name = "Example Package",
  resource_heb_name = "משאב לדוגמה",
  size = 1024,
  last_modified = Sys.time(),
  created = Sys.time()
)
as_gidi_resource(resource_data)
}
