% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gidigov.R
\name{data_gov_autocomplete_req}
\alias{data_gov_autocomplete_req}
\title{Create autocomplete request for data.gov.il}
\usage{
data_gov_autocomplete_req(q, req_type = c("organization", "dataset"))
}
\arguments{
\item{q}{Search query string}

\item{req_type}{Character string specifying type ("organization" or "dataset")}
}
\value{
An httr2 request object
}
\description{
Create autocomplete request for data.gov.il
}
