% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gidigov.R
\name{get_resource_info_by_name}
\alias{get_resource_info_by_name}
\title{Get resource information by name}
\usage{
get_resource_info_by_name(database_name)
}
\arguments{
\item{database_name}{Character string of database name}
}
\value{
data.table of resource information
}
\description{
Get resource information by name
}
