% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gidigov.R
\name{get_resource_name}
\alias{get_resource_name}
\title{Get a Resource's Display Name}
\usage{
get_resource_name(resource_id)
}
\arguments{
\item{resource_id}{Character string of a single resource ID.}
}
\value{
Character string -- the resource's display name.
}
\description{
Convenience wrapper around [get_resource_info()] that returns only the
resource name.
}
\seealso{
Other resource info: 
\code{\link{get_package_info}()},
\code{\link{get_resource_id_by_name}()},
\code{\link{get_resource_info}()},
\code{\link{get_resource_info_by_name}()}
}
\concept{resource info}
