% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gidi_obj_class.R
\name{gidi_obj_class}
\alias{gidi_obj_class}
\title{Base classes for gidi objects}
\description{
This file defines the base classes and methods for gidi objects:
- gidi_obj: Base class for individual gidi objects
- gidi_objs_list: Base class for lists of gidi objects
}
