% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gidi_org.R
\name{gidi_org_full_info}
\alias{gidi_org_full_info}
\title{Extended organization class with additional metrics}
\usage{
gidi_org_full_info(
  id = character(0),
  eng_name = character(0),
  heb_name = character(0),
  packages = integer(0),
  resources = integer(0)
)
}
\description{
Extends gidi_org with package and resource information
}
\note{
Inherits all properties from gidi_org
}
