% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gidi_search.R
\name{gidi_search_org}
\alias{gidi_search_org}
\title{Search for organizations in data.gov.il}
\usage{
gidi_search_org(q, as_data.frame = FALSE)
}
\arguments{
\item{q}{Search query string}

\item{as_data.frame}{Boolean, if TRUE returns results as data frame}
}
\value{
List or data frame of matching organizations
}
\description{
Search for organizations in data.gov.il
}
